+++ src/imio/dashboard/browser/views.py
@@ -5,11 +5,11 @@ import pkg_resources
 from plone import api
 from Products.Five.browser import BrowserView
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
-
 from collective.eeafaceted.collectionwidget.browser.views import RenderTermView as BaseRenderTermView
 from collective.eeafaceted.collectionwidget.widgets.widget import CollectionWidget
 
 from imio.dashboard.adapters import CURRENT_CRITERION
+from imio.dashboard.interfaces import IDashboardCollection
 from imio.dashboard.utils import getCollectionLinkCriterion
 
 
@@ -22,6 +22,8 @@ class RenderTermView(BaseRenderTermView):
 
     def display_number_of_items(self):
         """Display number of items in the collection."""
+        if not IDashboardCollection.providedBy(self.context):
+            return False
         return self.context.getShowNumberOfItems() or False
 
 
@@ -31,6 +33,8 @@ class RenderTermPortletView(BaseRenderTermView):
 
     def display_number_of_items(self):
         """Display number of items in the collection."""
+        if not IDashboardCollection.providedBy(self.context):
+            return False
         return self.context.getShowNumberOfItems() or False
 
     def __call__(self, term, category, widget):
@@ -58,7 +62,8 @@ class JSONCollectionsCount(BrowserView):
         for category in voc.itervalues():
             for term in category['collections']:
                 collection = api.content.get(UID=term.token)
-                if collection.getShowNumberOfItems():
+                if IDashboardCollection.providedBy(collection) \
+                        and collection.getShowNumberOfItems():
                     view = collection.unrestrictedTraverse(
                         '@@render_collection_widget_term_portlet')
                     info.append({